refactor: separate draft.js logic
